Extract caching code to separate files
